requirements.txt
PyQt6
PyMuPDF
Pytesseract
Pillow
Python-docx
